fixup! update the benchmark
